fix: use int8_t in table

+++ sim/predictor.h
@@ -7,11 +7,21 @@
 
 #define UINT16 unsigned short int
 
+/////////////// HARDWARE BUDGET ////////////////
+// Budget: 32KB
+//
+// GHR: 59 bits
+// Table: 512 * (59 + 1) * 8 bits
+//
+// 59 bits + 512 * 60 * 8 bits ≈ 30 KB
+////////////////////////////////////////////////
+
 class PREDICTOR
 {
 private:
   bitset<59> ghr;
-  int table[512][60];
+  // Use int8_t instead of int to fit more perceptrons under budget
+  int8_t table[512][60];
   int y;
 
 public:
